requirements.txt
Django
psycopg2
ptvsd
gunicorn
docutils
ipython
djangorestframework
pillow
